# Add optional FLAC compression