@mewmew 2025-07-08
> I'm open to feedback or questions, or if more tests are needed. I have no personal need for this, just thought it would be fun to try to enhance things a bit.

A labour born out of love and passion is the best <3

Really happy to see you implementing the compression for FLAC @braheezy! This has been a long awaited feature.

My brother and I are on summer holidays, going to the Swedish summer house today to have a Sauna : )

So, while we deeply appreciate the PR. Just giving a heads up that the review may take a while, as we are in vacaion-mode.

Joyous summer wishes,
Henry & Robin