@mewmew 2025-07-11
Really wonderful job with implementing encoding using Fixed prediction of samples!

I've merged the PR to main and done some stylistic changes.

The only issue I've identified is related to when trying to analyze FLAC files that have fewer than 2 samples in a subframe.

In such cases, the `sf.RiceSubframe` field is never initialized, since `analyzeFixed` has an early return before allocating the `&frame.RiceSubframe{}` struct.

```bash
$ go test ./...
--- FAIL: TestEncode (0.09s)
    --- FAIL: TestEncode/testdata/191885.flac (0.03s)
panic: runtime error: invalid memory address or nil pointer dereference [recovered]
	panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x10 pc=0x5342fc]

goroutine 18 [running]:
testing.tRunner.func1.2({0x56f280, 0x6deb00})
	/usr/lib/go/src/testing/testing.go:1734 +0x21c
testing.tRunner.func1()
	/usr/lib/go/src/testing/testing.go:1737 +0x35e
panic({0x56f280?, 0x6deb00?})
	/usr/lib/go/src/runtime/panic.go:792 +0x132
github.com/mewkiz/flac.analyzeSubframe(0xc000070620, 0x10)
	/home/u/life/projects/github.com/mewkiz/flac/analysis_fixed.go:175 +0xbc
github.com/mewkiz/flac.(*Encoder).WriteFrame(0xc00014e0f0, 0xc0000705b0)
	/home/u/life/projects/github.com/mewkiz/flac/encode_frame.go:93 +0x8e6
github.com/mewkiz/flac_test.TestEncode.func1(0xc000082380)
	/home/u/life/projects/github.com/mewkiz/flac/enc_test.go:150 +0x331
testing.tRunner(0xc000082380, 0xc000110198)
	/usr/lib/go/src/testing/testing.go:1792 +0xf4
created by testing.(*T).Run in goroutine 5
	/usr/lib/go/src/testing/testing.go:1851 +0x413
FAIL	github.com/mewkiz/flac	0.097s
ok  	github.com/mewkiz/flac/frame	(cached)
ok  	github.com/mewkiz/flac/internal/bits	(cached)
ok  	github.com/mewkiz/flac/internal/bufseekio	(cached)
?   	github.com/mewkiz/flac/internal/hashutil	[no test files]
ok  	github.com/mewkiz/flac/internal/hashutil/crc16	(cached)
ok  	github.com/mewkiz/flac/internal/hashutil/crc8	(cached)
?   	github.com/mewkiz/flac/internal/ioutilx	[no test files]
?   	github.com/mewkiz/flac/internal/utf8	[no test files]
ok  	github.com/mewkiz/flac/meta	(cached)
FAIL
```